// web/src/api/generated/model/auditEntryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AuditAction } from './auditAction';
import type { AuditChange } from './auditChange';

export interface AuditEntryDto {
  id: string;
  at: string;
  userId: string;
  action: AuditAction;
  entityName: string;
  entityLabelEn: string;
  entityLabelAr: string;
  entityId: string;
  /** @nullable */
  reference: string | null;
  changes: AuditChange[];
}